+++ src/main/java/com/example/socialmediafeed/domain/post/repository/PostRepository.java
@@ -1,7 +1,11 @@
 package com.example.socialmediafeed.domain.post.repository;
 
 import com.example.socialmediafeed.domain.post.entity.Post;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
+
+    Page<Post>  findByContentContainingIgnoreCase(String hashtag, Pageable pageable);
 }
